+++ bakery/views/__init__.py
@@ -1,5 +1,8 @@
 from .base import (
+    Buildable400View,
+    Buildable403View,
     Buildable404View,
+    Buildable500View,
     BuildableMixin,
     BuildableRedirectView,
     BuildableTemplateView,
@@ -14,7 +17,10 @@ from .detail import BuildableDetailView
 from .list import BuildableListView
 
 __all__ = (
+    "Buildable400View",
+    "Buildable403View",
     "Buildable404View",
+    "Buildable500View",
     "BuildableArchiveIndexView",
     "BuildableDayArchiveView",
     "BuildableDetailView",